chore: try to make eslint happy
