+++ packages/codemate-plugin/plugins/privilege-group/lib.ts
@@ -1,4 +1,4 @@
-import { Err, NotFoundError } from "hydrooj";
+import { Err, NotFoundError } from 'hydrooj';
 
 export const GroupNotFoundError = Err('GroupNotFoundError', NotFoundError);
 export const ActivationCodeNotFoundError = Err('ActivationCodeNotFoundError', NotFoundError);
